+++ packages/orq-rc/src/models/operations/runagent.ts
@@ -3975,7 +3975,7 @@ export const Tools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   Tools
 > = z.object({
-  id: z.string().default("01KWH6BYSX30VHQDNPV0WSMTY5"),
+  id: z.string().default("01KWHAAG7VAMGEYSGCMPQ3YF6W"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() => AgentToolInputRunSchema$outboundSchema),